Moved the util directory under tools
