+++ hed/tools/__init__.py
@@ -23,7 +23,9 @@ from .bids.bids_tabular_dictionary import BidsTabularDictionary
 from .bids.bids_tabular_file import BidsTabularFile
 from .bids.bids_tabular_summary import BidsTabularSummary
 
-from hed.tools.analysis.tabular_reports import report_diffs
+from .analysis.tabular_reports import report_diffs
 
-from hed.tools.util.hed_logger import HedLogger
-from hed.tools.util.io_util import get_file_list, parse_bids_filename
+from .util.hed_logger import HedLogger
+from .util.data_util import get_new_dataframe, get_value_dict, replace_values, reorder_columns
+from .util.io_util import check_filename, generate_filename, extract_suffix_path, get_file_list, make_path
+from .util.io_util import get_dir_dictionary, get_file_list, get_path_components, parse_bids_filename
