+++ hed/tools/analysis/hed_type_factors.py
