+++ hed/tools/bids/bids_tabular_summary.py
@@ -1,6 +1,6 @@
 
 from hed.errors import HedFileError
-from hed.tools.util import get_new_dataframe
+from hed.tools.util.data_util import get_new_dataframe
 from hed.tools.analysis.annotation_util import generate_sidecar_entry
