+++ hed/tools/util/__init__.py
@@ -1,5 +1,3 @@
 """ Data and file handling utilities."""
 
-from .data_util import get_new_dataframe, get_value_dict, replace_values, reorder_columns
-from .io_util import check_filename, generate_filename, extract_suffix_path, get_file_list, make_path
-from .io_util import get_dir_dictionary, get_file_list, get_path_components, parse_bids_filename
+
