+++ tests/tools/analysis/test_annotation_util.py
@@ -9,7 +9,7 @@ from hed.tools import BidsTabularSummary, check_df_columns, df_to_hed, extract_t
 from hed.tools.analysis.annotation_util import _find_last_pos, _find_first_pos, \
     _flatten_cat_col, _flatten_val_col, _get_value_entry, trim_back, trim_front, _tag_list_to_str, _update_cat_dict, \
     generate_sidecar_entry
-from hed.tools.util import get_file_list
+from hed.tools import get_file_list
 from hed.validator import HedValidator
